RpcClientProvider now also implements IClient.
