+++ src/ChainSafe.Gaming/Web3/Evm/JsonRpc/RpcClientProvider.cs
@@ -5,13 +5,14 @@ using ChainSafe.Gaming.Web3.Analytics;
 using ChainSafe.Gaming.Web3.Core;
 using ChainSafe.Gaming.Web3.Environment;
 using Nethereum.Hex.HexTypes;
+using Nethereum.JsonRpc.Client;
 using Nethereum.JsonRpc.Client.RpcMessages;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
 namespace ChainSafe.Gaming.Evm.Providers
 {
-    public class RpcClientProvider : IRpcProvider, ILifecycleParticipant
+    public class RpcClientProvider : ClientBase, IRpcProvider, ILifecycleParticipant
     {
         private readonly RpcClientConfig config;
         private readonly Web3Environment environment;
@@ -100,19 +101,9 @@ namespace ChainSafe.Gaming.Evm.Providers
 
             try
             {
-                var httpClient = environment.HttpClient;
                 var request = new RpcRequestMessage(Guid.NewGuid().ToString(), method, parameters);
-                var response =
-                    (await httpClient.Post<RpcRequestMessage, RpcResponseMessage>(config.RpcNodeUrl, request))
-                    .AssertSuccess();
 
-                if (response.HasError)
-                {
-                    var error = response.Error;
-                    var errorMessage =
-                        $"RPC returned error for \"{method}\": {error.Code} {error.Message} {error.Data}";
-                    throw new Web3Exception(errorMessage);
-                }
+                var response = await SendAsync(request);
 
                 var serializer = JsonSerializer.Create();
                 return serializer.Deserialize<T>(new JTokenReader(response.Result))!;
@@ -133,6 +124,27 @@ namespace ChainSafe.Gaming.Evm.Providers
                     PackageName = "io.chainsafe.web3-unity",
                 });
             }
+        }
+
+        protected override async Task<RpcResponseMessage> SendAsync(RpcRequestMessage request, string route = null)
+        {
+            string body = JsonConvert.SerializeObject(request);
+
+            return await SendAsyncInternally<RpcResponseMessage>(body);
+        }
+
+        protected override async Task<RpcResponseMessage[]> SendAsync(RpcRequestMessage[] requests)
+        {
+            string body = JsonConvert.SerializeObject(requests);
+
+            return await SendAsyncInternally<RpcResponseMessage[]>(body);
+        }
+
+        private async Task<T> SendAsyncInternally<T>(string body)
+        {
+            var result = await environment.HttpClient.PostRaw(config.RpcNodeUrl, body, "application/json");
+
+            return JsonConvert.DeserializeObject<T>(result.Response);
         }
     }
 }
